Move secret hash protocol to sub directories
In preparation for adding the remaining one shot communication protocols
move the secret hash protocol to sub directories of `behaviour/` and
`protocol/`.
